real-time bpm + latency

+++ Note.cpp
@@ -1,7 +1,7 @@
 #include "Note.h"
 #include <iostream>
 
-Note::Note(int p, int d, int v)
+Note::Note(int p, double d, int v)
 	: pitch{p}, duration{d}, velocity{v}{
 	// std::cout << "Note ctor !!!" << std::endl;
 
@@ -10,7 +10,7 @@ Note::Note(const Note &source)
 	: pitch{source.pitch}, duration{source.duration}, velocity{source.velocity}{
 }
 
-Rest::Rest(int d) : Note(0, d, 0){
+Rest::Rest(double d) : Note(0, d, 0){
 	// std::cout << "Rest ctor" << std::endl;
 }
